Split out `match_same_arms`
